pyproject.toml
[project]
name = "fastapi-resume"
version = "0.0.1"
description = "FastAPI-powered resume generator: expose your CV as a JSON REST API"
readme = "README.md"
maintainers = [
  {name = "Nick Kelly", email = "nialexanderke@proton.me"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.116.1",
    "orjson==3.11.0",
    "pyyaml==6.0.2",
    "typer==0.16.0",
    "uvicorn[standard]==0.32.1",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed"
]
keywords=[
    "Python",
    "resume",
    "project directory",
    "package",
    "packaging",
    "FastAPI",
    "Typer",
    "CLI"
]

[dependency-groups]
dev = [
    "black==25.1.0",
    "isort==6.0.1",
    "pytest==8.3.3",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "mypy==1.16.1",
    "ruff==0.12.2",
    "types-pyyaml==6.0.12.20250516",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
fast-resume = "fastapi_resume.cli:app"

[project.urls]
Homepage = "https://github.com/nickatnight/fastapi-resume"
Issues = "https://github.com/nickatnight/fastapi-resume/issues"

[tool.uv]
required-version = ">=0.7.18"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["fastapi_resume"]  # Replace with your actual package name
exclude = ["docs*", "tests*", "templates*"] 

[tool.isort]
multi_line_output = 3
lines_after_imports = 2
line_length = 88
include_trailing_comma = true
sections = 'FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER'

[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
(
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)
'''

[tool.ruff]
line-length = 88
exclude = [".git", "__pycache__", ".mypy_cache", ".pytest_cache", "migrations", ".venv"]
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    # "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
lint.ignore = [
    "B904",
    "B006",
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "F403",  # import * only allowed to support wildcard imports
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.codespell]
ignore-words-list = 'selectin,THIRDPARTY'

[tool.coverage.report]

[tool.coverage.run]
omit = ['*tests/*']

[tool.mypy]
# --strict
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
implicit_reexport = false
strict_equality = true
# --strict end

[[tool.mypy.overrides]]
module = ["typer.*"]
ignore_missing_imports = true
